lambda/requirements.txt
line-bot-sdk
openai